+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added option `--tdb` to `odoo-helper odoo recompute` command
 - Added new option `--no-backup` to `odoo-helper install reinstall-odoo`
 - On module migration automatically replace `phantom_js(` to `browser_js(`.
+- Added new option `--if-not-exists` to `odoo-helper db create` command
 
 ### Changed